fix geo_modif persistance #14313

+++ src/main/java/fr/microtec/geo2/persistance/repository/common/GeoModifRepository.java
@@ -1,10 +1,12 @@
 package fr.microtec.geo2.persistance.repository.common;
 
+import java.math.BigDecimal;
+
 import org.springframework.stereotype.Repository;
 
 import fr.microtec.geo2.persistance.entity.common.GeoModification;
 import fr.microtec.geo2.persistance.repository.GeoRepository;
 
 @Repository
-public interface GeoModifRepository extends GeoRepository<GeoModification, Integer> {
+public interface GeoModifRepository extends GeoRepository<GeoModification, BigDecimal> {
 }
